requirements.txt
-f /home/somnium/wheels
attrs==19.3.0
backcall==0.1.0
chardet==3.0.4
Click==7.0
decorator==4.4.1
et-xmlfile==1.0.1
Flask==1.1.1
Flask-Caching==1.8.0
Flask-SocketIO==4.2.1
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.1
MarkupSafe==1.1.1
nbformat==4.4.0
numpy==1.17.4
openpyxl==3.0.2
pandas==0.25.3
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pkg-resources==0.0.0
plotly==4.3.0
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.5.1
pyrsistent==0.15.6
python-dateutil==2.8.1
python-engineio==3.10.0
python-socketio==4.4.0
pytz==2019.3
pyzmq==18.1.1
retrying==1.3.3
RPi.GPIO==0.7.0
schedule==0.6.0
six==1.13.0
tornado==6.0.3
traitlets==4.3.3
wcwidth==0.1.7
Werkzeug==0.16.0
xlrd==1.2.0
